refactor: Rename `lib/clientUtils.js`
