+++ lib/app-uids.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { getPlatformClientWithAccessKey } = require('./clientUtils');
+const { getPlatformClientWithAccessKey } = require('./client-utils');
 
 module.exports = async function (orgName, appName) {
   const sdk = await getPlatformClientWithAccessKey(orgName);
