+++ lib/app-uids.test.js
@@ -6,7 +6,7 @@ const sinon = require('sinon');
 
 const getAppStub = sinon.stub().resolves({ appUid: 'AUID', tenantUid: 'OUID' });
 const appUids = proxyquire('./app-uids', {
-  './clientUtils': {
+  './client-utils': {
     getPlatformClientWithAccessKey: () => ({
       apps: {
         get: getAppStub,
