+++ lib/client-utils.js
