+++ lib/client-utils.test.js
@@ -21,12 +21,12 @@ const platformClientStub = {
   },
 };
 
-describe('lib/clientUtils.test.js', () => {
+describe('lib/client-utils.test.js', () => {
   const orgName = 'testorg';
 
   describe('getOrCreateAccessKeyForOrg', () => {
     it('returns SERVERLESS_ACCESS_KEY from env if set', async () => {
-      const { getOrCreateAccessKeyForOrg } = proxyquire('./clientUtils', {
+      const { getOrCreateAccessKeyForOrg } = proxyquire('./client-utils', {
         '@serverless/utils/config': {
           getLoggedInUser: () => null,
         },
@@ -42,7 +42,7 @@ describe('lib/clientUtils.test.js', () => {
     });
 
     it('throws an error if user does not exist', async () => {
-      const { getOrCreateAccessKeyForOrg } = proxyquire('./clientUtils', {
+      const { getOrCreateAccessKeyForOrg } = proxyquire('./client-utils', {
         '@serverless/utils/config': {
           getLoggedInUser: () => null,
         },
@@ -56,7 +56,7 @@ describe('lib/clientUtils.test.js', () => {
     it('returns an existing access key from config', async () => {
       const accessKeyFromConfig = 'existingaccesskey';
 
-      const { getOrCreateAccessKeyForOrg } = proxyquire('./clientUtils', {
+      const { getOrCreateAccessKeyForOrg } = proxyquire('./client-utils', {
         '@serverless/utils/config': {
           getLoggedInUser: () => ({
             userId: 'someuserid',
@@ -72,7 +72,7 @@ describe('lib/clientUtils.test.js', () => {
     });
 
     it('throws if it could not create an access key when user does not have idToken', async () => {
-      const { getOrCreateAccessKeyForOrg } = proxyquire('./clientUtils', {
+      const { getOrCreateAccessKeyForOrg } = proxyquire('./client-utils', {
         '@serverless/utils/config': {
           getLoggedInUser: () => ({
             userId: 'someuserid',
@@ -87,7 +87,7 @@ describe('lib/clientUtils.test.js', () => {
 
     it('succesfully returns and persists created access key', async () => {
       const configUtilsSetStub = sinon.stub();
-      const { getOrCreateAccessKeyForOrg } = proxyquire('./clientUtils', {
+      const { getOrCreateAccessKeyForOrg } = proxyquire('./client-utils', {
         '@serverless/utils/config': {
           getLoggedInUser: () => ({ userId: 'userid', idToken: 'sometoken', username: 'user' }),
           set: configUtilsSetStub,
