+++ lib/dashboard.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const open = require('open');
-const { getPlatformClientWithAccessKey } = require('./clientUtils');
+const { getPlatformClientWithAccessKey } = require('./client-utils');
 const isAuthenticated = require('./isAuthenticated');
 const log = require('./log');
