+++ lib/dashboard.test.js
@@ -21,7 +21,7 @@ const ServerlessSDKMock = class ServerlessSDK {
 };
 
 const { dashboardHandler, getDashboardProvidersUrl } = proxyquire('./dashboard', {
-  './clientUtils': {
+  './client-utils': {
     getPlatformClientWithAccessKey: async () => new ServerlessSDKMock(),
   },
   './isAuthenticated': () => isAuthenticated,
