+++ lib/deployProfile.js
@@ -4,7 +4,7 @@ const { style } = require('@serverless/utils/log');
 const log = require('./log');
 const { serviceSlug, instanceSlug } = require('./utils');
 const { getDashboardUrl } = require('./dashboard');
-const { getPlatformClientWithAccessKey } = require('./clientUtils');
+const { getPlatformClientWithAccessKey } = require('./client-utils');
 
 module.exports.configureDeployProfile = async (ctx) => {
   let deploymentProfile;
