+++ lib/deployProfile.test.js
@@ -21,7 +21,7 @@ describe('configureDeployProfile', () => {
 
     ({ configureDeployProfile } = proxyquire('./deployProfile', {
       '@serverless/platform-client': { ServerlessSDK },
-      './clientUtils': {
+      './client-utils': {
         getPlatformClientWithAccessKey: () => ({
           deploymentProfiles: {
             get: deployProfileGetStub,
