+++ lib/deployment/deployment.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const _ = require('lodash');
-const { getPlatformClientWithAccessKey } = require('../clientUtils');
+const { getPlatformClientWithAccessKey } = require('../client-utils');
 const platformClientVersion = require('@serverless/platform-client/package').version;
 
 module.exports = class {
