+++ lib/deployment/deployment.test.js
@@ -9,7 +9,7 @@ describe('lib/deployment/deployment.test.js', () => {
   const frameworkDeploymentCreateStub = sinon.stub().resolves();
 
   const Deployment = proxyquire('./deployment', {
-    '../clientUtils': {
+    '../client-utils': {
       getPlatformClientWithAccessKey: async () => ({
         frameworkDeployments: {
           create: frameworkDeploymentCreateStub,
