+++ lib/logsCollection.js
@@ -17,7 +17,7 @@ const {
   LAMBDA_FILTER_PATTERN,
 } = require('./utils');
 
-const { getPlatformClientWithAccessKey } = require('./clientUtils');
+const { getPlatformClientWithAccessKey } = require('./client-utils');
 
 const formatRegionName = (regionName) => {
   return regionName
