+++ lib/logsCollection.test.js
@@ -13,7 +13,7 @@ describe('logsCollection', () => {
   before(() => {
     getLogDestinationStub = sinon.stub().resolves({ destinationArn: 'arn:logdest' });
     logsCollection = proxyquire('./logsCollection', {
-      './clientUtils': {
+      './client-utils': {
         getPlatformClientWithAccessKey: () => ({
           logDestinations: {
             getOrCreate: getLogDestinationStub,
