+++ lib/removeDestination.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { getPlatformClientWithAccessKey } = require('./clientUtils');
+const { getPlatformClientWithAccessKey } = require('./client-utils');
 
 module.exports = async (ctx) => {
   if (
