+++ lib/removeDestination.test.js
@@ -10,7 +10,7 @@ describe('removeDestination', () => {
   before(() => {
     removeLogDestinationSpy = sinon.spy();
     removeDestination = proxyquire('./removeDestination', {
-      './clientUtils': {
+      './client-utils': {
         getPlatformClientWithAccessKey: () => ({
           logDestinations: {
             remove: removeLogDestinationSpy,
