+++ lib/resolveOutputs.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { getPlatformClientWithAccessKey } = require('./clientUtils');
+const { getPlatformClientWithAccessKey } = require('./client-utils');
 
 module.exports = async ({ org, app, service, stage, region }) => {
   const sdk = await getPlatformClientWithAccessKey(org);
