+++ lib/resolveParams.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const memoizee = require('memoizee');
 const { serviceSlug, instanceSlug } = require('./utils');
-const { getPlatformClientWithAccessKey } = require('./clientUtils');
+const { getPlatformClientWithAccessKey } = require('./client-utils');
 const isAuthenticated = require('./isAuthenticated');
 
 const mapDashboardParamType = (paramType) => {
