+++ lib/resolveProviderCredentials.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { serviceSlug, instanceSlug } = require('./utils');
-const { getPlatformClientWithAccessKey } = require('./clientUtils');
+const { getPlatformClientWithAccessKey } = require('./client-utils');
 
 module.exports = async ({ configuration: { org, app, service }, region, stage }) => {
   const sdk = await getPlatformClientWithAccessKey(org);
