+++ lib/resolveProviderCredentials.test.js
@@ -45,7 +45,7 @@ describe('lib/resolveProviderCredentials.test.js', () => {
 
   it('resolves to null if no providers found', async () => {
     const resolveProviderCredentials = proxyquire('./resolveProviderCredentials', {
-      './clientUtils': {
+      './client-utils': {
         getPlatformClientWithAccessKey: async () => ({
           ...mockedSdk,
           getProvidersByOrgServiceInstance: async () => {
@@ -70,7 +70,7 @@ describe('lib/resolveProviderCredentials.test.js', () => {
 
   it('resolves to null org was not found', async () => {
     const resolveProviderCredentials = proxyquire('./resolveProviderCredentials', {
-      './clientUtils': {
+      './client-utils': {
         getPlatformClientWithAccessKey: async () => ({
           organizations: {
             get: async () => {
@@ -96,7 +96,7 @@ describe('lib/resolveProviderCredentials.test.js', () => {
 
   it('succesfully resolves credentials', async () => {
     const resolveProviderCredentials = proxyquire('./resolveProviderCredentials', {
-      './clientUtils': {
+      './client-utils': {
         getPlatformClientWithAccessKey: async () => mockedSdk,
       },
     });
