+++ lib/wrap.js
@@ -11,7 +11,7 @@ const path = require('path');
 const _ = require('lodash');
 const JSZip = require('jszip');
 const log = require('./log');
-const { getOrCreateAccessKeyForOrg } = require('./clientUtils');
+const { getOrCreateAccessKeyForOrg } = require('./client-utils');
 const { addTree, writeZip } = require('./zipTree');
 const { version } = require('../package.json');
