+++ lib/wrap.test.js
@@ -463,7 +463,7 @@ except Exception as error:
         'jszip': JSZip,
         './zipTree': { addTree, writeZip },
         'fs-extra': fs,
-        './clientUtils': {
+        './client-utils': {
           getOrCreateAccessKeyForOrg: () => 'accesskey',
         },
       });
